refactor. 검색 관련 컴포넌트 Product로 이동
